Create exporter.go and move Exporter struct with its funcs into it
